Handle angle of collision edge case
WLOG, it's possible for a left-turn-left-side collision point to have a
negative y value. In this case, the angle appears to be large since it's
range constrained to [0, 2PI). The math in place actually still works as
long as we constrain the calculated difference between the angles. I
have added comments documenting the edge case.